requirements.txt
# Core dependencies
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Async & HTTP
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# Caching (optional)
redis>=4.5.0
hiredis>=2.0.0

# Progress & CLI
tqdm>=4.65.0
click>=8.1.0
rich>=13.0.0

# Logging & Monitoring
loguru>=0.7.0
prometheus-client>=0.17.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
isort>=5.12.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0

# Utilities
python-dateutil>=2.8.0
regex>=2023.6.0